Moved test out of `pastey` crate
This is done as the internal crate was required to be published that only contains a basic macro. The other option here could be include that macro in `pastey` which is available on a feature, but I didn't went with that approach as I want to do some isolation for the test cases and they might grew in future.